fixed Python/Setup.ps1 when no .venv exists

+++ Python/Setup.ps1
@@ -1,4 +1,6 @@
-rm .venv -Recurse -Force -Confirm:$false
+if(Test-Path -Path ./.venv){
+	rm .venv -Recurse -Force -Confirm:$false
+}
 python -m venv .venv
 .\.venv\Scripts\Activate.ps1
 pip install -r requirements.txt
